+++ ftauth/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Bump `json_annotation` and `json_serializable` dependencies
+- Fix `Config.toJson`
+
 ## 0.1.0
 
 - Initial release